GaussianPooling class

This class represents a model with pooled Gaussian components, inheriting from GaussianStaticBorrowing. The power parameter is set to 1.

Super classes

Model -> GaussianConjugate -> GaussianStaticBorrowing -> GaussianPooling

Public fields

method

Method name

Methods

Public methods

GaussianPooling$new()

GaussianPooling$clone()

Inherited methods

Model$calibrate_for_design()

Model$check_data()

Model$create()

Model$empirical_bayes_update()

Model$estimate_bayesian_operating_characteristics()

Model$estimate_frequentist_operating_characteristics()

Model$hypothesis_space_transformation()

Model$inference()

Model$inference_cache_scope()

Model$plot_pdfs()

Model$plot_posterior_pdf()

Model$plot_prior_pdf()

Model$posterior_beta_mixture()

Model$posterior_ess()

Model$posterior_quantile()

Model$print_model_summary()

Model$prior_elir_ess()

Model$simulation_for_given_treatment_effect()

Model$test_decision()

GaussianConjugate$credible_interval()

GaussianConjugate$posterior_cdf()

GaussianConjugate$posterior_mean()

GaussianConjugate$posterior_median()

GaussianConjugate$posterior_moments()

GaussianConjugate$posterior_pdf()

GaussianConjugate$posterior_to_RBesT()

GaussianConjugate$posterior_variance()

GaussianConjugate$prior_cdf()

GaussianConjugate$prior_pdf()

GaussianConjugate$prior_to_RBesT()

GaussianConjugate$sample_posterior()

GaussianConjugate$sample_prior()

GaussianConjugate$vectorised_posterior_parameters()

GaussianConjugate$vectorised_replicate_inference()

GaussianStaticBorrowing$summary_rows()

GaussianStaticBorrowing$vectorised_prior_variance()

GaussianPooling$new()

Usage

GaussianPooling$new(prior)

Arguments

prior

Prior

Returns

A Model object.

GaussianPooling$clone()

The objects of this class are cloneable with this method.

Usage

GaussianPooling$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

NA
#> [1] NA