BinomialCPP Class

A class representing a Binomial Conditional Power Prior model.

Super classes

Model -> MCMCModel -> BinomialCPP

Public fields

power_parameter

The power parameter for the model

method

Method name

stan_prior

Stan prior model

stan_prior_code

Stan code for the prior

Methods

Public methods

BinomialCPP$new()

BinomialCPP$prepare_data()

BinomialCPP$draw_mcmc_prior()

BinomialCPP$clone()

Inherited methods

Model$check_data()

Model$create()

Model$empirical_bayes_update()

Model$estimate_bayesian_operating_characteristics()

Model$estimate_frequentist_operating_characteristics()

Model$inference_cache_scope()

Model$plot_pdfs()

Model$plot_posterior_pdf()

Model$plot_prior_pdf()

Model$posterior_beta_mixture()

Model$posterior_mean()

Model$posterior_moments()

Model$posterior_quantile()

Model$posterior_to_RBesT()

Model$prior_ESS()

Model$prior_elir_ess()

Model$prior_to_RBesT()

Model$prior_treatment_benefit()

Model$simulation_for_given_treatment_effect()

Model$test_decision()

Model$vectorised_replicate_inference()

MCMCModel$check_mcmc_config()

MCMCModel$compute_posterior_parameters()

MCMCModel$credible_interval()

MCMCModel$inference()

MCMCModel$posterior_cdf()

MCMCModel$posterior_ess()

MCMCModel$posterior_median()

MCMCModel$posterior_pdf()

MCMCModel$prior_cdf()

MCMCModel$prior_pdf()

MCMCModel$sample_posterior()

MCMCModel$sample_prior()

BinomialCPP$new()

Initialize an instance of the BinomialCPP class.

Usage

BinomialCPP$new(prior, mcmc_config)

Arguments

prior

The prior object

mcmc_config

The MCMC configuration parameters

BinomialCPP$prepare_data()

Prepare data for the model

Usage

BinomialCPP$prepare_data(target_data)

Arguments

target_data

The target study data

Returns

A list of prepared data

BinomialCPP$draw_mcmc_prior()

Sample from the prior using Stan

Usage

BinomialCPP$draw_mcmc_prior()

BinomialCPP$clone()

The objects of this class are cloneable with this method.

Usage

BinomialCPP$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.